Agrego funcion para escribir en el espacio de usuario y arreglo un poco el ida y vuelta en el modulo CPU

+++ memoria/src/marcos.c
@@ -136,7 +136,7 @@ void asignar_marcos_a_tabla(t_tabla_nivel* tabla, int pid) {
 }
 
 int obtener_pid_por_dir_fisica(uint32_t direccion_fisica){
-    int nro_marco = ((int) direccion_fisica) / TAM_PAGINA;
+    int nro_marco = direccion_fisica / TAM_PAGINA;
 
     t_marco* marco = obtener_marco_por_nro_marco(nro_marco);
